+++ app/src/main/java/com/onelio/connectu/Containers/NotaData.java
@@ -52,7 +52,7 @@ public class NotaData {
         return nota;
     }
 
-    public void setNota(float nota) {
+    public void setNota(double nota) {
         this.nota = nota;
     }
